Update jquery-migrate to version 3.6.0
Replace the bundled jquery-migrate-3.4.1.js with the official 3.6.0
uncompressed build and update the references in web-ui-wro-sources.xml
and base-layout-cssjs-loader.xsl accordingly.

jquery-migrate 3.6.0 targets jQuery >=3 <4, so it stays compatible
with the bundled jQuery 3.7.1. The jquery-migrate-mute.js helper and
the script load order are unchanged.
