+++ web-ui/src/main/resources/WEB-INF/classes/web-ui-wro-sources.xml
@@ -36,7 +36,7 @@
     <jsSource webappPath="/catalog/lib/modernizr.js" minimize="false"/>
     <jsSource webappPath="/catalog/lib/jquery-3.7.1.js" minimize="true"/>
     <jsSource webappPath="/catalog/lib/jquery-migrate-mute.js" minimize="false"/>
-    <jsSource webappPath="/catalog/lib/jquery-migrate-3.4.1.js" minimize="false"/>
+    <jsSource webappPath="/catalog/lib/jquery-migrate-3.6.0.js" minimize="false"/>
     <jsSource webappPath="/catalog/lib/moment-with-locales.min.js" minimize="false"/>
     <jsSource webappPath="/catalog/lib/moment-timezone-with-data-1970-2030.min.js" minimize="false" />
     <jsSource webappPath="/catalog/lib/angular/angular.min.js" minimize="false"/>
@@ -126,7 +126,7 @@
     <jsSource webappPath="/catalog/lib/modernizr.js" minimize="false"/>
     <jsSource webappPath="/catalog/lib/jquery-3.7.1.js" minimize="true"/>
     <jsSource webappPath="/catalog/lib/jquery-migrate-mute.js" minimize="false"/>
-    <jsSource webappPath="/catalog/lib/jquery-migrate-3.4.1.js" minimize="false"/>
+    <jsSource webappPath="/catalog/lib/jquery-migrate-3.6.0.js" minimize="false"/>
     <jsSource webappPath="/catalog/lib/moment-with-locales.min.js" minimize="false"/>
     <jsSource webappPath="/catalog/lib/moment-timezone-with-data-1970-2030.min.js" minimize="false" />
     <jsSource webappPath="/catalog/lib/angular/angular.min.js" minimize="false"/>
